+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-06-26
+
+### Added
+
+- Add option to specify a schema name [#44](https://github.com/policy-design-lab/data-ingestion/issues/44)
+
+### Fixed
+
+- Bug in practice names when there are nested parentheses in the raw
+  data [#51](https://github.com/policy-design-lab/data-ingestion/issues/51)
+- Create database was not passing in db_port when connecting to
+  database [#65](https://github.com/policy-design-lab/data-ingestion/issues/65)
+
+### Changed
+
+- Update code to include Title-II data from years 2014 to
+  2023 [#54](https://github.com/policy-design-lab/data-ingestion/issues/54)
+- Update code to include Crop Insurance data from years 2014 to
+  2023 [#63](https://github.com/policy-design-lab/data-ingestion/issues/63)
+
 ## [0.4.1] - 2024-11-05
 
 ### Fixed
@@ -64,6 +84,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - This CHANGELOG file.
 - Feature to parse and ingest Title II data. [#7](https://github.com/policy-design-lab/data-ingestion/issues/7)
+
+[0.5.0]: https://github.com/policy-design-lab/data-ingestion/compare/0.4.1...0.5.0
 
 [0.4.1]: https://github.com/policy-design-lab/data-ingestion/compare/0.4.0...0.4.1